Corrected DTO types. Some `Ints` needed to tbe `Doubles` and the Data class from JSON file did not pick it up.

+++ app/src/main/java/com/example/andrdoid_clean_architecture/data/remote/dto/BeerDto.kt
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName
 data class BeerDto(
     val abv: Double,
     @SerializedName("attenuation_level")
-    val attenuationLevel: Int,
+    val attenuationLevel: Double,
     @SerializedName("boil_volume")
     val boilVolume: BoilVolumeDto,
     @SerializedName("brewers_tips")
@@ -20,7 +20,7 @@ data class BeerDto(
     val firstBrewed: String,
     @SerializedName("food_pairing")
     val foodPairing: List<String>,
-    val ibu: Int,
+    val ibu: Double,
     val id: Int,
     @SerializedName("image_url")
     val imageUrl: String,
@@ -28,12 +28,12 @@ data class BeerDto(
     val method: MethodDto,
     val name: String,
     val ph: Double,
-    val srm: Int,
+    val srm: Double,
     val tagline: String,
     @SerializedName("target_fg")
     val targetFg: Int,
     @SerializedName("target_og")
-    val targetOg: Int,
+    val targetOg: Double,
     val volume: VolumeDto
 )
